Create Array_Permutation_1_to_N.cpp
A sequence of N integers is called a permutation if it contains all integers from 1 to N exactly once. This code is to find if the array taken from the user is a permutation of numbers from 1 to N, in C++. The idea used here is that : if the size of the set if equal to N then the array represents a permutation else it doesn’t.